fix: plan定数をクライアント安全なモジュールに分離
plan.ts がサーバー専用モジュール（env-validation）に依存しており、
クライアントコンポーネント（SupportPlanSection）から import すると
モジュール評価時に CSRF_TOKEN_SALT バリデーションがクライアント側で
実行されクラッシュしていた。

定数・型を plan-constants.ts に分離し、クライアントからはそちらを参照。
